refactor(repo): split classes into their own files
